requirements.txt

pandas>=2.0.0
pywhatkit>=5.4
phonenumbers>=8.13.0
pyyaml>=6.0
tqdm>=4.65.0
openpyxl>=3.1.0
requests>=2.31.0
click>=8.1.0
